Types file has been renamed to actionsTypes.js
