+++ MusicCatalog/index.js
@@ -3,6 +3,7 @@ import { Navigation } from 'react-native-navigation';
 import { RootNavigation } from './src/navigation/RootNavigation';
 
 if (__DEV__) {
+    
 import('./ReactotronConfig').then(() => console.log('Reactotron Configured'));
 }
