+++ MusicCatalog/src/constants/actionsTypes.js
